Refactor : useInfinitePlaylist, query와 service로 분리
